rename routes to pages
