+++ homepage/src/App.js
@@ -1,11 +1,11 @@
 import { useEffect, useState} from 'react';
-import Home from './routes/Home';
-import DesignPatterns from './routes/DesignPatterns';
-import Designers from './routes/Designers';
-import Developers from './routes/Developers';
-import Components from './routes/Components';
-import Foundations from './routes/Foundations';
-import Support from './routes/Support';
+import Home from './pages/Home';
+import DesignPatterns from './pages/DesignPatterns';
+import Designers from './pages/Designers';
+import Developers from './pages/Developers';
+import Components from './pages/Components';
+import Foundations from './pages/Foundations';
+import Support from './pages/Support';
 import ClipLoader from 'react-spinners/ClipLoader';
 import fetchContentfulAPI from './imageRendering/imageRendering';
 import useImage from "./imageRendering/ImageContext";
