+++ homepage/src/pages/DesignPatterns.jsx
